chore: clean up paseo runtime
