+++ Cargo.toml
@@ -120,9 +120,10 @@ scale-info = { version = "2.11.6", default-features = false, features = [
 ] }
 
 # External dependencies
-log = { version = "0.4.17", default-features = false }
 serde = { version = "1.0", default-features = false, features = ["derive"] }
 serde_json = { version = "1.0", default-features = false }
+tracing = { version = "0.1.41", default-features = false }
+tracing-subscriber = { version = "=0.3.19" }
 
 # Async/HTTP (for provider node)
 axum = "0.7"
