+++ storage-interfaces/file-system/client/Cargo.toml
@@ -22,7 +22,7 @@ subxt = "0.37"
 subxt-signer = "0.37"
 
 # Utilities
-log = { workspace = true }
+tracing = { workspace = true }
 thiserror = "2.0"
 hex = "0.4"
